+++ cmake/fbuffer.cmake
@@ -13,10 +13,13 @@ add_executable(
     ${DARWIN_SOURCES}
     samples/fbuffer/BufferTask.cpp samples/fbuffer/BufferTask.hpp
     samples/fbuffer/BufferThreadManager.cpp samples/fbuffer/BufferThreadManager.hpp
+    samples/fbuffer/BufferThread.cpp samples/fbuffer/BufferThread.hpp
     samples/fbuffer/Generator.cpp samples/fbuffer/Generator.hpp
     samples/fbuffer/Connectors/AConnector.cpp samples/fbuffer/Connectors/AConnector.hpp
     samples/fbuffer/Connectors/fAnomalyConnector.cpp samples/fbuffer/Connectors/fAnomalyConnector.hpp
     samples/fbuffer/Connectors/fSofaConnector.cpp samples/fbuffer/Connectors/fSofaConnector.hpp
+    toolkit/AThreadManager.cpp toolkit/AThreadManager.hpp
+    toolkit/AThread.cpp toolkit/AThread.hpp
 )
 
 target_link_libraries(
